[feature] complete associations request handler
- also renamed routeHandlers folder to requestHandlers
